Ensure network configuration chain IDs are hex strings
We have seen reports of users with an invalid NetworkController state,
namely, that some network configurations have decimal chain IDs or are
filed under properties that are decimal chain IDs. This breaks
`wallet_getPermissions` in the dapp connection flow.

This commit adds a migration that corrects both of these cases so that
any decimal chain IDs are now hex strings.